refactor: split up article generator file
